fix: remove weekend spend key from expenses (#3154)
