+++ src/app/core/enums/platform/v1/expense-transaction-status.enum.ts
@@ -1,0 +1,4 @@
+export enum ExpenseTransactionStatus {
+  PENDING = 'PENDING',
+  POSTED = 'POSTED',
+}
