+++ src/app/core/models/platform/v1/accounting-export-summary.model.ts
@@ -1,0 +1,6 @@
+export interface AccountingExportSummary {
+  state: string;
+  error_type: string;
+  url: string;
+  tpa_id: string;
+}
