+++ src/app/core/models/platform/v1/expense-missing-mandatory-fields.model.ts
@@ -1,0 +1,6 @@
+export interface ExpenseMissingMandatoryFields {
+  expense_field_ids: number[];
+  amount: boolean;
+  currency: boolean;
+  receipt: boolean;
+}
