+++ src/app/core/models/platform/v1/expense-rule-data.model.ts
@@ -1,0 +1,11 @@
+import { NameValuePair } from '../../name-value-pair.model';
+
+export interface ExpenseRuleData {
+  merchant: string;
+  is_billable: boolean;
+  purpose: string;
+  category_id: number;
+  project_id: number;
+  cost_center_id: number;
+  custom_fields: NameValuePair[];
+}
