+++ src/app/core/models/platform/v1/violation.model.ts
@@ -1,0 +1,4 @@
+export interface Violation {
+  policy_rule_description: string;
+  policy_rule_id: string;
+}
