+++ src/app/fyle/add-edit-expense/add-edit-expense-6.spec.ts
@@ -73,7 +73,6 @@ import {
 } from 'src/app/core/test-data/accounts.service.spec.data';
 import { expectedProjectsResponse } from 'src/app/core/test-data/projects.spec.data';
 import { AddEditExpensePage } from './add-edit-expense.page';
-import { TransactionStatus } from 'src/app/core/models/platform/v1/expense.model';
 import { TransactionStatusInfoPopoverComponent } from 'src/app/shared/components/transaction-status-info-popover/transaction-status-info-popover.component';
 import { ExpensesService } from 'src/app/core/services/platform/v1/spender/expenses.service';
 import { AdvanceWalletsService } from 'src/app/core/services/platform/v1/spender/advance-wallets.service';
@@ -87,6 +86,7 @@ import { matchedCCTransactionData, matchedCCTransactionData2 } from 'src/app/cor
 import { ccTransactionData, ccTransactionData1 } from 'src/app/core/mock-data/cc-transaction.data';
 import { ccTransactionResponseData } from 'src/app/core/mock-data/corporate-card-transaction-response.data';
 import { cloneDeep } from 'lodash';
+import { ExpenseTransactionStatus } from 'src/app/core/enums/platform/v1/expense-transaction-status.enum';
 
 export function TestCases6(getTestBed) {
   describe('AddEditExpensePage-6', () => {
@@ -1210,14 +1210,14 @@ export function TestCases6(getTestBed) {
       const popoverSpy = jasmine.createSpyObj('HTMLIonPopoverElement', ['present']);
       popoverController.create.and.resolveTo(popoverSpy);
 
-      component.openTransactionStatusInfoModal(TransactionStatus.PENDING);
+      component.openTransactionStatusInfoModal(ExpenseTransactionStatus.PENDING);
 
       tick();
 
       expect(popoverController.create).toHaveBeenCalledOnceWith({
         component: TransactionStatusInfoPopoverComponent,
         componentProps: {
-          transactionStatus: TransactionStatus.PENDING,
+          transactionStatus: ExpenseTransactionStatus.PENDING,
         },
         cssClass: 'fy-dialog-popover',
       });
