+++ src/app/shared/components/transaction-status-info-popover/transaction-status-info-popover.component.html
@@ -18,7 +18,7 @@
 
 <ion-content class="transaction-status-info-popover__content">
   <div class="transaction-status-info-popover__body" data-testid="content">
-    <ng-container *ngIf="transactionStatus === TransactionStatus.PENDING; else posted"
+    <ng-container *ngIf="transactionStatus === 'PENDING'; else posted"
       >Your transaction status is Pending until your bank processes the transaction.</ng-container
     >
